Move Data Union classes to dataunion package
In order to reduce cyclic dependencies between packages.
