+++ src/main/java/com/streamr/client/dataunion/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Data Union package has dependencies to:
+ * <ul>
+ *     <li>org.web3j:crypto</li>
+ *     <li>org.web3k:core</li>
+ *     <li>org.web3k:utils</li>
+ *     <li>org.slf4j:slf4j-api</li>
+ * </ul>
+ */
+package com.streamr.client.dataunion;
